Add user ownership to video and handle related errors
- Add `requireToken` middleware to create, update, delete actions for video
- Manage error handling with ownership requirements
- Update videos.json with more seed data

+++ app/routes/video_routes.js
@@ -1,7 +1,7 @@
 // Express docs: http://expressjs.com/en/api.html
 const express = require('express')
 // Passport docs: http://www.passportjs.org/docs/
-// const passport = require('passport')
+const passport = require('passport')
 
 // pull in Mongoose model for videos
 const Video = require('../models/video')
@@ -14,6 +14,7 @@ const customErrors = require('../../lib/custom_errors')
 const handle404 = customErrors.handle404
 // we'll use this function to send 401 when a user tries to modify a resource
 // that's owned by someone else
+const requireOwnership = customErrors.requireOwnership
 
 // this is middleware that will remove blank fields from `req.body`, e.g.
 // { video: { title: '', text: 'foo' } } -> { video: { text: 'foo' } }
@@ -21,7 +22,7 @@ const removeBlanks = require('../../lib/remove_blank_fields')
 // passing this as a second argument to `router.<verb>` will make it
 // so that a token MUST be passed for that route to be available
 // it will also set `req.user`
-// const requireToken = passport.authenticate('bearer', { session: false })
+const requireToken = passport.authenticate('bearer', { session: false })
 
 // instantiate a router (mini app that only handles routes)
 const router = express.Router()
@@ -56,7 +57,9 @@ router.get('/videos/:id', (req, res, next) => {
 
 // CREATE
 // POST /videos
-router.post('/videos', (req, res, next) => {
+router.post('/videos', requireToken, (req, res, next) => {
+  req.body.video.owner = req.user.id
+
   Video.create(req.body.video)
     // respond to succesful `create` with status 201 and JSON of new "video"
     .then(video => {
@@ -70,12 +73,18 @@ router.post('/videos', (req, res, next) => {
 
 // UPDATE
 // PATCH /videos/5a7db6c74d55bc51bdf39793
-router.patch('/videos/:id', removeBlanks, (req, res, next) => {
+router.patch('/videos/:id', requireToken, removeBlanks, (req, res, next) => {
   // if the client attempts to change the `owner` property by including a new
   // owner, prevent that by deleting that key/value pair
+  delete req.body.video.owner
+
   Video.findById(req.params.id)
     .then(handle404)
     .then(video => {
+      // pass the 'req' object and the Mongoose record to `requireOnwer`
+      // it will throw an error if the current user isn't the owner
+      requireOwnership(req, video)
+
       // pass the result of Mongoose's `.update` to the next `.then`
       return video.updateOne(req.body.video)
     })
@@ -87,7 +96,8 @@ router.patch('/videos/:id', removeBlanks, (req, res, next) => {
 
 // DESTROY
 // DELETE /videos/5a7db6c74d55bc51bdf39793
-router.delete('/videos/:id', (req, res, next) => {
+router.delete('/videos/:id', requireToken, (req, res, next) => {
+  // handle admin privilege
   // if (req.user.role === 'admin') {
   //   Video.findById(req.params.id)
   //     .then(handle404)
@@ -96,16 +106,19 @@ router.delete('/videos/:id', (req, res, next) => {
   //     })
   //     .then(() => res.sendStatus(204))
   //     .catch(next)
-  // }
+  // } else {
   Video.findById(req.params.id)
     .then(handle404)
     .then(video => {
+      // throw an error if current user doesn't own video
+      requireOwnership(req, video)
       video.deleteOne()
     })
     // send back 204 and no content if the deletion succeeded
     .then(() => res.sendStatus(204))
     // if an error occurs, pass it to the handler
     .catch(next)
+  // }
 })
 
 module.exports = router
